07- SIMVA as built images + fix simva checkout code + profiling and logs folders

+++ docker-stacks/bin/purge-folder-contents.sh
@@ -5,6 +5,8 @@ set -euo pipefail
 folders_to_remove=$@
 
 for folder in ${folders_to_remove}; do
-  echo "Removing content of folder $folder"
-  find "$folder" -mindepth 1 -maxdepth 1 ! -name '.gitignore'  -exec rm -rf {} +
+  if [[ -e "${folder}" ]]; then
+    echo "Removing content of folder $folder"
+    find "$folder" -mindepth 1 -maxdepth 1 ! -name '.gitignore'  -exec rm -rf {} +
+  fi
 done;
